Implement switching between different views in design tab
